Изменена структура redux файлов
