+++ src/components/Filters/index.js
@@ -1,6 +1,6 @@
 import React from "react";
 import FilterLink from "../../containers/FilterLink";
-import {VisibilityFilters} from "../../actions";
+import {VisibilityFilters} from "../../modules/visibilityFilter";
 
 export const Filter = () => (
     <div>
