+++ src/containers/AddTodo/index.js
@@ -1,6 +1,6 @@
 import React from "react";
 import {connect} from "react-redux";
-import {addTodo} from "../../actions";
+import {addTodo} from "../../modules/todos";
 import styles from "./AddTodo.module.css";
 
 const AddTodo = ({dispatch}) => {
