+++ src/index.js
@@ -4,7 +4,7 @@ import './index.module.css';
 import App from './components/App';
 import * as serviceWorker from './serviceWorker';
 import {createStore} from "redux";
-import rootReducer from "./redusers";
+import rootReducer from "./modules";
 import {Provider} from "react-redux";
 
 const store = createStore(rootReducer);
