+++ src/modules/index.js
@@ -1,0 +1,8 @@
+import { combineReducers } from 'redux';
+import {todos} from './todos';
+import {visibilityFilter} from './visibilityFilter';
+
+export default combineReducers({
+    todos,
+    visibilityFilter
+})
