+++ src/modules/todos/actions.js
@@ -11,11 +11,3 @@ export const addTodo = createAction("ADD_TODO", (text) => ({
 export const toggleTodo = createAction("TOGGLE_TODO", id => id);
 
 export const deleteTodo = createAction("DELETE_TODO", id => id);
-
-export const setVisibilityFilter = createAction("SET_VISIBILITY_FILTER", filter => filter);
-
-export const VisibilityFilters = {
-    SHOW_ALL: "SHOW_ALL",
-    SHOW_ACTIVE: "SHOW_ACTIVE",
-    SHOW_DONE: "SHOW_DONE",
-}
