+++ src/modules/todos/reducer.js
@@ -1,12 +1,10 @@
-import {combineReducers} from "redux";
 import {handleActions} from "redux-actions";
 import {createSelector} from "reselect";
 import {addTodo, deleteTodo, toggleTodo} from "./actions";
-import {setVisibilityFilter, VisibilityFilters} from "./actions";
 
 const initialState = [];
 
-const todos = handleActions({
+export const todos = handleActions({
         [addTodo]: (state, action) => [...state, action.payload],
         [deleteTodo]: (state, action) => (state.filter((todo) => {
             return !(todo.id === action.payload);
@@ -24,15 +22,3 @@ const todos = handleActions({
             })
         )
 }, initialState);
-
-const visibilityFilter = handleActions({
-    [setVisibilityFilter]: (state, action) => action.payload,
-}, VisibilityFilters.SHOW_ALL);
-
-export const getTodos = state => state.todos;
-export const getVisibilityFilter = state => state.visibilityFilter;
-
-export default combineReducers({
-    todos,
-    visibilityFilter
-})
