+++ src/modules/todos/selectors.js
@@ -1,0 +1,1 @@
+export const getTodos = state => state.todos;
