+++ src/modules/visibilityFilter/index.js
@@ -1,0 +1,3 @@
+export { visibilityFilter } from './reducer';
+export * from './selectors';
+export * from './actions';
