+++ src/modules/visibilityFilter/reducer.js
@@ -1,0 +1,6 @@
+import {handleActions} from "redux-actions";
+import {setVisibilityFilter, VisibilityFilters} from "./actions";
+
+export const visibilityFilter = handleActions({
+    [setVisibilityFilter]: (state, action) => action.payload,
+}, VisibilityFilters.SHOW_ALL);
