+++ src/modules/visibilityFilter/selectors.js
@@ -1,0 +1,1 @@
+export const getVisibilityFilter = state => state.visibilityFilter;
